leo: Patch libmmcamera2_iface_modules to ignore CDS requests.
CDS is broken on kernel 4.9. As default the following properties are set:
persist.camera.CDS=off
persist.camera.video.CDS=off
They are meant as default and doesn't prevent user space apps to set
CDS modes, which results in green and disorted images. This is seen
on an popular messenger W****A** when capturing images in a very dark
environment.
On camera HAL3 the CDS requests are transferred via functions of the
camera.msm8953.so library to the functions within libmmcamera2_iface_modules.
Since it's hard to shim the right functions of the camera.msm8953.so blob,
patching the libmmcamera2_iface_modules library is easier to suppress
CDS setting.

This is done by patching the relevant value assignment where not
the value is taken of the parameter, but a static value 0x0, which
means the CDS mode is set to OFF.

[SebaUbuntu] Correct address for vince